RPC-API: add getinfo endpoint.
This commit creates a new endpoint /getinfo which currently returns only
the version of Joinmarket (JM_CORE_VERSION) running in the backend
joinmarketd. In future commits the returned dict may add more
information, for example the current block height of the connected
Bitcoin Core node.

Additionally, this commit fixes the response type of the
/rescanblockchain endpoint as defined in the OpenAPI spec.