Remove CLI option --restart from tumbler
Prior to this commit, the --restart option was still present in the help
documentation of tumbler.py, however this functionality has been removed
from the algorithm (restarting now just means "running again", since the
algorithm now functions the same way independent of the location of the
coins in mixdepths). Moreover, running with this option could lead to
crash conditions.
After this commit, the --restart option is removed from the help text,
and no longer executed in code. Further, a slight change is made to the
wording of the tumblerguide.md documentation, section "Restarting", to
make the situation clearer.