Delay !auth sending until all makers receive !fill
Prior to this commit, in scenarios of very big latency differentials
between maker bots, it was possible for one maker to complete the
conversation up to !ioauth and broadcast the commitment, before another
maker had even received the !fill message. This causes incorrect local
blacklisting of commitments for that (and any other slow) maker.
After this commit, the taker does not continue Phase1 with the !auth
message, until either (a) all of the makers have sent their !pubkey
message (indicating that they have already seen the commitment in the
!fill message), or (b) a timeout of 60 seconds (after which any maker
that has not sent the !pubkey is ignored for the rest of the
conversation).